+++ include/DrawingManager.h
@@ -47,10 +47,11 @@ public:
 	sf::Text* addText(FontType font, std::string string, const sf::Color& col, size_t size, const sf::Vector2f& pos);
 	void addText(sf::Text* text);
 	void addSprite(std::string filePath, const sf::Vector2f& pos = { 0, 0 }, const sf::Vector2f& scale = { 1.0, 1.0 });
-	void addRect(const sf::Vector2f& size, const sf::Vector2f& pos, const sf::Color& col = { 255, 0, 0 });
-	void addButton(const sf::Vector2f& pos, std::string textureFile, std::function<void()> onClick, bool isOwned = true);
-	void addButton(Button* btn, bool isOwned = true);
+	sf::RectangleShape* addRect(const sf::Vector2f& size, const sf::Vector2f& pos, const sf::Color& col = { 255, 0, 0 });
+	Button* addButton(const sf::Vector2f& pos, std::string textureFile, std::function<void()> onClick, bool isOwned = true);
+	Button* addButton(Button* btn, bool isOwned = true);
 	void addToggleButton(const sf::Vector2f& pos, std::string textureFile, std::string textureOnFile = "none", std::function<void()> onClick = {}, bool isOwned = true, bool isToggleSfx = true, bool setOn = false);
+	void addLine(const sf::Vector2f pos1, const sf::Vector2f pos2, const sf::Color col = {0, 255, 255}, float thickness = 4.f);
 	void draw(sf::RenderWindow* window) const;
 
 	// Buttons